Move template and git-related code into a tmt.utils submodule
