+++ docs/scripts/generate-template-filters.py
@@ -3,7 +3,8 @@
 import sys
 import textwrap
 
-from tmt.utils.templates import TEMPLATE_FILTERS, Path, render_template_file
+from tmt.utils import Path
+from tmt.utils.templates import TEMPLATE_FILTERS, render_template_file
 
 HELP = textwrap.dedent("""
 Usage: generate-template-filters.py <TEMPLATE-PATH> <OUTPUT-PATH>
