+++ tests/integration/test_data/test_nitrate/NitrateExport.test_export_forced_validation.yaml
@@ -326,7 +326,7 @@ nitrate.xmlrpc_driver:
         summary: tmt /validation
         tag:
         - 10700
-tmt.utils:
+tmt.utils.git:
   check_git_url:
   - metadata:
       guess_type: Simple
