+++ tmt/utils/__init__.py
@@ -4085,172 +4085,6 @@ def create_file(
     echo(message)
 
 
-# Avoid multiple subprocess calls for the same url
-@functools.cache
-def check_git_url(url: str) -> str:
-    """ Check that a remote git url is accessible """
-    try:
-        log.debug(f"Check git url '{url}'.")
-        subprocess.check_call(
-            ["git", "ls-remote", "--heads", url],
-            stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL,
-            env={"GIT_ASKPASS": "echo", "GIT_TERMINAL_PROMPT": "0"})
-        return url
-    except subprocess.CalledProcessError:
-        raise GitUrlError(f"Unable to contact remote git via '{url}'.")
-
-
-PUBLIC_GIT_URL_PATTERNS: list[tuple[str, str]] = [
-    # Gitlab on private namespace is synced to pkgs.devel.redhat.com
-    # old: https://gitlab.com/redhat/rhel/tests/bash
-    # old: git@gitlab.com:redhat/rhel/tests/bash
-    # new: https://pkgs.devel.redhat.com/git/tests/bash
-    (
-        r'(?:git@|https://)gitlab.com[:/]redhat/rhel(/.+)',
-        r'https://pkgs.devel.redhat.com/git\1'
-        ),
-
-    # GitHub, GitLab
-    # old: git@github.com:teemtee/tmt.git
-    # new: https://github.com/teemtee/tmt.git
-    (
-        r'git@(.*):(.*)',
-        r'https://\1/\2'
-        ),
-
-    # RHEL packages
-    # old: git+ssh://psplicha@pkgs.devel.redhat.com/tests/bash
-    # old: ssh://psplicha@pkgs.devel.redhat.com/tests/bash
-    # old: ssh://pkgs.devel.redhat.com/tests/bash
-    # new: https://pkgs.devel.redhat.com/git/tests/bash
-    (
-        r'(git\+)?ssh://(\w+@)?(pkgs\.devel\.redhat\.com)/(.*)',
-        r'https://\3/git/\4'
-        ),
-
-    # Fedora packages, Pagure
-    # old: git+ssh://psss@pkgs.fedoraproject.org/tests/shell
-    # old: ssh://psss@pkgs.fedoraproject.org/tests/shell
-    # new: https://pkgs.fedoraproject.org/tests/shell
-    (
-        r'(git\+)?ssh://(\w+@)?([^/]*)/(.*)',
-        r'https://\3/\4'
-        )
-    ]
-
-
-def public_git_url(url: str) -> str:
-    """
-    Convert a git url into a public format.
-
-    :param url: an URL to convert.
-    :returns: URL that is publicly accessible without authentication,
-        or the original URL if no applicable conversion was found.
-    """
-    return rewrite_git_url(url, PUBLIC_GIT_URL_PATTERNS)
-
-
-def rewrite_git_url(url: str, patterns: list[tuple[str, str]]) -> str:
-    """
-    Rewrite git url based on supplied patterns
-
-    :param url: an URL to modify
-    :param patterns: List of patterns to try in order
-    :returns: Modified url or the original one if no pattern was be applied.
-    """
-    for pattern, replacement in patterns:
-        public_url = re.sub(pattern, replacement, url)
-
-        # If we got different string, `pattern` matched the URL and
-        # `replacement` made its changes - we got our hit!
-        if public_url != url:
-            return public_url
-
-    # Otherwise return unmodified
-    return url
-
-
-# Environment variable prefixes
-INJECT_CREDENTIALS_URL_PREFIX = 'TMT_GIT_CREDENTIALS_URL_'
-INJECT_CREDENTIALS_VALUE_PREFIX = 'TMT_GIT_CREDENTIALS_VALUE_'
-
-
-def inject_auth_git_url(url: str) -> str:
-    """
-    Inject username or token to the git url
-
-    :param url: original git repo url
-    :returns: URL with injected authentication based on pattern from the environment
-        or unmodified URL
-    """
-    # Try all environment variables sorted by their name
-    for name, value in sorted(os.environ.items(), key=lambda x: x[0]):
-        # First one which matches url is taken into the account
-        if name.startswith(INJECT_CREDENTIALS_URL_PREFIX) and re.search(value, url):
-            unique_suffix = name[len(INJECT_CREDENTIALS_URL_PREFIX):]
-            variable_with_value = f'{INJECT_CREDENTIALS_VALUE_PREFIX}{unique_suffix}'
-            # Get credentials value
-            try:
-                creds = os.environ[variable_with_value]
-            except KeyError:
-                raise GitUrlError(
-                    f'Missing "{variable_with_value}" variable with credentials for "{url}"')
-            # Return original url if credentials is an empty value
-            if not creds:
-                return url
-            # Finally inject credentials into the url and return it
-            return re.sub(r'([^/]+://)([^/]+)', rf'\1{creds}@\2', url)
-    # Otherwise return unmodified
-    return url
-
-
-CLONABLE_GIT_URL_PATTERNS: list[tuple[str, str]] = [
-    # git:// protocol is not possible for r/o access
-    # old: git://pkgs.devel.redhat.com/tests/bash
-    # new: https://pkgs.devel.redhat.com/git/tests/bash
-    (
-        r'git://(pkgs\.devel\.redhat\.com)/(.*)',
-        r'https://\1/git/\2'
-        ),
-    ]
-
-
-def clonable_git_url(url: str) -> str:
-    """ Modify the git repo url so it can be cloned """
-    url = rewrite_git_url(url, CLONABLE_GIT_URL_PATTERNS)
-    return inject_auth_git_url(url)
-
-
-def web_git_url(url: str, ref: str, path: Optional[Path] = None) -> str:
-    """
-    Convert a public git url into a clickable web url format
-
-    Compose a clickable link from git url, ref and path to file
-    for the most common git servers.
-    """
-    if path:
-        path = Path(urllib.parse.quote_plus(str(path), safe="/"))
-
-    # Special handling for pkgs.devel (ref at the end)
-    if 'pkgs.devel' in url:
-        url = url.replace('git://', 'https://').replace('.com', '.com/cgit')
-        url += '/tree'
-        if path:
-            url += str(path)
-        url += f'?h={ref}'
-        return url
-
-    # GitHub & GitLab
-    if any(server in url for server in ['github', 'gitlab']):
-        url = url.replace('.git', '').rstrip('/')
-        url += f'/tree/{ref}'
-
-    if path:
-        url += str(path)
-
-    return url
-
-
 @functools.cache
 def fmf_id(
         *,
@@ -4282,6 +4116,8 @@ def fmf_id(
     except ValueError:
         remote_name = 'origin'
     remote = run(Command("git", "config", "--get", f"remote.{remote_name}.url"))
+
+    from tmt.utils.git import default_branch, git_root, public_git_url
     fmf_id.url = public_git_url(remote) if remote else None
 
     # Construct path (if different from git root)
@@ -4444,220 +4280,6 @@ class retry_session(contextlib.AbstractContextManager):  # type: ignore[type-arg
 def remove_color(text: str) -> str:
     """ Remove ansi color sequences from the string """
     return re.sub(r'\x1B(?:[@-Z\\-_]|\[[0-?]*[ -/]*[@-~])', '', text)
-
-
-def git_hash(*, directory: Path, logger: tmt.log.Logger) -> Optional[str]:
-    """
-    Return short hash of current HEAD in the git repo in directory.
-
-    :param directory: path to a local git repository.
-    :param logger: used for logging.
-    :returns: short hash as string
-    """
-    cmd = Command("git", "rev-parse", "--short", "HEAD")
-    result = cmd.run(cwd=directory, logger=logger)
-
-    if result.stdout is None:
-        raise RunError(message="No output from 'git' when looking for the hash of HEAD.",
-                       command=cmd,
-                       returncode=0,
-                       stderr=result.stderr)
-
-    return result.stdout.strip()
-
-
-def git_root(*, fmf_root: Path, logger: tmt.log.Logger) -> Optional[Path]:
-    """
-    Find a path to the root of git repository containing an fmf root.
-
-    :param fmf_root: path to an fmf root that is supposedly in a git repository.
-    :param logger: used for logging.
-    :returns: path to the git repository root, if fmf root lies in one,
-        or ``None``.
-    """
-
-    try:
-        result = Command("git", "rev-parse", "--show-toplevel").run(cwd=fmf_root, logger=logger)
-
-        if result.stdout is None:
-            return None
-
-        return Path(result.stdout.strip())
-
-    except RunError:
-        # Always return an empty string in case 'git' command is run in a non-git repo
-        return None
-
-
-def git_add(*, path: Path, logger: tmt.log.Logger) -> None:
-    """
-    Add path to the git index.
-
-    :param path: path to add to the git index.
-    :param logger: used for logging.
-    """
-    path = path.resolve()
-
-    try:
-        Command("git", "add", path).run(cwd=path.parent, logger=logger)
-
-    except RunError as error:
-        raise GeneralError(f"Failed to add path '{path}' to git index.") from error
-
-
-def git_ignore(*, root: Path, logger: tmt.log.Logger) -> list[Path]:
-    """
-    Collect effective paths ignored by git.
-
-    :param root: path to the root of git repository.
-    :param logger: used for logging.
-    :returns: list of actual paths that would be ignored by git based on
-        its ``.gitignore`` files. If a whole directory is to be ignored,
-        it is listed as a directory path, not listing its content.
-    """
-
-    output = Command(
-        'git',
-        'ls-files',
-        # Consider standard git exclusion files
-        '--exclude-standard',
-        # List untracked files matching exclusion patterns
-        '-oi',
-        # If a whole directory is to be ignored, list only its name with a trailing slash
-        '--directory') \
-        .run(cwd=root, logger=logger)
-
-    return [Path(line.strip()) for line in output.stdout.splitlines()] if output.stdout else []
-
-
-def default_branch(
-        *,
-        repository: Path,
-        remote: str = 'origin',
-        logger: tmt.log.Logger) -> Optional[str]:
-    """ Detect default branch from given local git repository """
-    # Make sure '.git' is present and it is a file or a directory
-    dot_git = repository / '.git'
-    if not dot_git.exists():
-        return None
-    if not dot_git.is_file() and not dot_git.is_dir():
-        return None
-
-    # Detect the original repository path if worktree is provided
-    if dot_git.is_file():
-        try:
-            result = Command("git", "rev-parse", "--path-format=absolute", "--git-common-dir").run(
-                cwd=repository,
-                logger=logger)
-        except RunError:
-            return None
-        if result.stdout is None:
-            return None
-        repository = Path(result.stdout.strip().replace("/.git", ""))
-
-    # Make sure the '.git/refs/remotes/{remote}' directory is present
-    git_remotes_dir = repository / f'.git/refs/remotes/{remote}'
-    if not git_remotes_dir.exists():
-        return None
-
-    # Make sure the HEAD reference is available
-    head = git_remotes_dir / 'HEAD'
-    if not head.exists():
-        try:
-            Command('git', 'remote', 'set-head', f'{remote}', '--auto').run(
-                cwd=repository,
-                logger=logger)
-        except BaseException:
-            return None
-
-    # The ref format is 'ref: refs/remotes/origin/main'
-    return head.read_text().strip().split('/')[-1]
-
-
-def validate_git_status(test: 'tmt.base.Test') -> tuple[bool, str]:
-    """
-    Validate that test has current metadata on fmf_id
-
-    Return a tuple (boolean, message) as the result of validation.
-
-    Checks that sources:
-    - all local changes are committed
-    - up to date on remote repository
-    - .fmf/version marking fmf root is committed as well
-
-    When all checks pass returns (True, '').
-    """
-    sources = [
-        *test.fmf_sources,
-        test.fmf_root / '.fmf' / 'version'
-        ]
-
-    # Use tmt's run instead of subprocess.run
-    run = Common(logger=test._logger).run
-
-    # Check for not committed metadata changes
-    cmd = Command(
-        'git',
-        'status', '--porcelain',
-        '--',
-        *[str(source) for source in sources]
-        )
-    try:
-        result = run(cmd, cwd=Path(test.node.root), join=True)
-    except RunError as error:
-        return (
-            False,
-            f"Failed to run git status: {error.stdout}"
-            )
-
-    not_committed = []
-    assert result.stdout is not None
-    for line in result.stdout.split('\n'):
-        if line:
-            # XY PATH or XY ORIG -> PATH. XY and PATH are separated by space
-            not_committed.append(line[3:])
-
-    if not_committed:
-        return (False, "Uncommitted changes in " + " ".join(not_committed))
-
-    # Check for not pushed changes
-    cmd = Command("git", "rev-parse", "--abbrev-ref", "--symbolic-full-name", "@{u}")
-    try:
-        result = run(cmd, cwd=Path(test.node.root))
-    except RunError as error:
-        return (
-            False,
-            f'Failed to get remote branch, error raised: "{error.stderr}"'
-            )
-
-    assert result.stdout is not None
-    remote_ref = result.stdout.strip()
-
-    cmd = Command(
-        'git',
-        'diff',
-        f'HEAD..{remote_ref}',
-        '--name-status',
-        '--',
-        *[str(source) for source in sources]
-        )
-    try:
-        result = run(cmd, cwd=Path(test.node.root))
-    except RunError as error:
-        return (
-            False,
-            f'Failed to diff against remote branch, error raised: "{error.stderr}"')
-
-    not_pushed = []
-    assert result.stdout is not None
-    for line in result.stdout.split('\n'):
-        if line:
-            _, path = line.strip().split('\t', maxsplit=2)
-            not_pushed.append(path)
-    if not_pushed:
-        return (False, "Not pushed changes in " + " ".join(not_pushed))
-
-    return (True, '')
 
 
 def generate_runs(
@@ -5305,81 +4927,6 @@ def distgit_download(
         target_dir.mkdir(exist_ok=True, parents=True)
         with open(target_dir / source_name, 'wb') as tarball:
             tarball.write(response.content)
-
-
-def git_clone(
-        *,
-        url: str,
-        destination: Path,
-        shallow: bool = False,
-        can_change: bool = True,
-        env: Optional[Environment] = None,
-        attempts: Optional[int] = None,
-        interval: Optional[int] = None,
-        timeout: Optional[int] = None,
-        logger: tmt.log.Logger) -> CommandOutput:
-    """
-    Clone git repository from provided url to the destination directory
-
-    :param url: Source URL of the git repository.
-    :param destination: Full path to the destination directory.
-    :param shallow: For ``shallow=True`` first try to clone repository
-        using ``--depth=1`` option. If not successful clone repo with
-        the whole history.
-    :param can_change: URL can be modified with hardcoded rules. Use
-        ``can_change=False`` to disable rewrite rules.
-    :param env: Environment provided to the ``git clone`` process.
-    :param attempts: Number of tries to call the function.
-    :param interval: Amount of seconds to wait before a new try.
-    :param timeout: Overall maximum time in seconds to clone the repo.
-    :param logger: A Logger instance to be used for logging.
-    :returns: Command output, bundled in a :py:class:`CommandOutput` tuple.
-    """
-
-    def clone_the_repo(
-            url: str,
-            destination: Path,
-            shallow: bool = False,
-            env: Optional[Environment] = None,
-            timeout: Optional[int] = None) -> CommandOutput:
-        """ Clone the repo, handle history depth """
-
-        depth = ['--depth=1'] if shallow else []
-        return Command('git', 'clone', *depth, url, destination).run(
-            cwd=Path('/'), env=env, timeout=timeout, logger=logger)
-
-    timeout = timeout or GIT_CLONE_TIMEOUT
-    attempts = attempts or GIT_CLONE_ATTEMPTS
-    interval = interval or GIT_CLONE_INTERVAL
-
-    # Update url only once
-    if can_change:
-        url = clonable_git_url(url)
-
-    # Do an extra shallow clone first
-    if shallow:
-        try:
-            return clone_the_repo(
-                shallow=True,
-                url=url,
-                destination=destination,
-                env=env,
-                timeout=timeout)
-        except RunError:
-            logger.debug(f"Shallow clone of '{url}' failed, let's try with the full history.")
-
-    # Finish with whatever number attempts requested (deep)
-    return retry(
-        func=clone_the_repo,
-        attempts=attempts,
-        interval=interval,
-        label=f"git clone {url} {destination}",
-        url=url,
-        destination=destination,
-        shallow=False,
-        env=env,
-        timeout=timeout,
-        logger=logger)
 
 
 # ignore[type-arg]: base class is a generic class, but we cannot list its parameter type, because
